// src/registry/models/ServiceAccountV1.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { CredentialsV1 } from './CredentialsV1';
export type ServiceAccountV1 = {
    created?: string;
    createdBy?: string;
    credentials?: CredentialsV1;
    description?: string;
    lastUpdatedBy?: string;
    name?: string;
    redHatAccountId?: string;
    tokenDate?: string;
};